♻️ Refactor: Partitioner 공통 사용 클래스로 수정
